// CMS/src/app/Modules/Users/model/user.model.ts
export class User {
    username: any;
    firstName: any;
    lastName: any;
    contact: any;
    email: any;
    password: any;
    confirmPassword: any;
    gender: any;
}
//nammade matte project edukkavo  ethu nammde uploadum ithaalunmnoi okke ollathu nammal full aayittu cheythekkane aano upload ulppede
// integrate cheyyathe nammalu indakkithu foinalaytt ithano nikku uploadillathe registration mathram aytt nammlu
// cheythille only registration athalle adhyam edutheaah athu thanne
//ihil illa 